adding: pieces available moves show in screen (with a lot of bugs) new make target rebuild and brun commands fixing build command

+++ include/GameLogic/PieceMoveGenerator.h
@@ -6,12 +6,12 @@
 #include "GameLogic/Board.h"
 #include "GameLogic/GameData.h"
 
-Bitboard bpawn_moves(Board board, uint64_t square);
-Bitboard wpawn_moves(Board board, uint64_t square);
-Bitboard rook_moves(Players player, Board board, uint64_t square);
-Bitboard bishop_moves(Players player, Board board, uint64_t square);
-Bitboard knight_moves(Players player, Board board, uint64_t square);
-Bitboard queen_moves(Players player, Board board, uint64_t square);
-Bitboard king_moves(Players player, Board board, uint64_t square);
+Bitboard bpawn_moves(Board board, short square);
+Bitboard wpawn_moves(Board board, short square);
+Bitboard rook_moves(Players player, Board board, short square);
+Bitboard bishop_moves(Players player, Board board, short square);
+Bitboard knight_moves(Players player, Board board, short square);
+Bitboard queen_moves(Players player, Board board, short square);
+Bitboard king_moves(Players player, Board board, short square);
 
 #endif
